requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1

# Telegram
python-telegram-bot==21.9

# Claude API
anthropic==0.42.0

# Monitoring
sentry-sdk[fastapi]==2.19.2

# Utils
python-dotenv==1.0.1
pydantic-settings==2.7.1
